backend/requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.1
django-cors-headers==4.3.1
fastapi==0.109.0
Flask==3.0.1
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
mysql-connector==2.2.9
mysql-connector-python==8.3.0
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.21.12
pyarrow==15.0.0
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.4
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
selenium==4.17.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.35.1
streamlit==1.30.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
undetected-chromedriver==2.0.0
urllib3==2.1.0
uvicorn==0.27.0.post1
validators==0.22.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
